Greatly simplify module structure
This reverts some of the changes made a couple commits ago to enable using a
custom EvalState. It turned out that this wasn't as useful as imagined and
everything necessary probably could be just implemented in the basic eval state
without too much loss of generality. This has the side effect of greatly
simplifying the module logic in general and hopefully make type signatures more
readable.