Big change to uuid logic
Random uuids are no longer generated for every astNode. Instead,
when a node needs to be stored in the uuidmap, a hash of the node's
contents is generated and used as the uuid.

This can't use Persist.to_string because there is special logic to
ignore the funcT.recur field in order to be able to use this scheme
and still support recursive functions.

TODO: handle collisions by having the uuidmap be id -> list astNodeT
and make the corresponding necessary changes to bot/Index.re